requirements.txt
pyhamcrest
allure-behave
allure-python-commons
selenium

